The function type constructor now has levity arguments.
Adam Gundry found this fix.

This is in line with the already applied fix to isFunCat
in 23de7c5e394e030113b1f5b10487f29491fae163.

With this patch, the golden tests now successfully build
on ghc-8.2.2, but there are 10 test suite failures where
the output is deviating from the expected output in a
subtle way.