Use a more composable form of `override`
When defining the overlay for the ConCat Haskell packages I've
provided a "static" argument of the form { overrides = hself: hsuper:
{...}} to the `override` function. As I've just learned this does not
compose well: it loses information of previous layers that used
`override`.

A better way is to provide a function as arguments to `override` which
takes any previous overrides and composes them with the provided
overlay.